refactor: split daemon config and error modules
